Fix boss.c's Random prototype (f64 -> f32), BossSpewCoins 90.23% -> 93.12%
boss.c declared `extern f64 Random(f64 a)` against the ml_fmath.c definition
`f32 Random(f32 scale)`, so all nine `Random(0.1f)` call sites widened the
argument and loaded the f64 0x3FB99999A0000000 where retail loads the f32
0x3DCCCCCD.  Nine wrong-width pool references, invisible to every arbiter.

Correcting the prototype also moves the score: real 137 -> 121, fresh fuzzy
90.2329% -> 93.1190%.  Game Code fuzzy 98.12% -> 98.13%; matched and the
STRICT/EQUIVALENT split (56.63 + 10.09) unchanged.

+++ src/game/boss/boss.c
@@ -89,7 +89,7 @@ extern void* CritterNewInst(int a, int b, void* obj);
 extern void* CopyMat4(void* a, void* b);
 extern void* PlaceItem(s32 a, s32 b, char* name, void* mat);
 extern void YawVec3(void* axis, f32* out, f32 angle);
-extern f64  Random(f64 a);
+extern f32  Random(f32 a);
 extern s32  StartFXSub(s32 type, f32* pos, u32 fla, u32 flb, f32 time);
 extern void SfxSetDamage(f32 damage, f32 radius, f32 delay, s32 idx, s32 type, s32 owner);
